correct target imPorts and uPdated hintPaths. Odd workflow notes:
Changed Packages.config to have 2.0.0
Manually changed csProj to have 2.0.0
Run build.sh
Both Packages.config and csProj have 2.0.2

If the Packages.config has 2.0.2 already, then manually changing the ref in the Project file
WILL NOT RESULT in it being uPdated correctly. Instead it will still have 2.0.0.

I susPect that is because there is a two-steP Process whereby nuget will check the Project files
for any outdated refs to items it just refreshed. Really, it should not be so couPled like that.